refactor: Separate image decoding/compression logic
